watchman/requirements.txt
kubernetes==23.6.0
ecs-logging==2.0.0
dynaconf==3.1.9
dnslib==0.9.20
dnspython==2.2.1
redis==4.3.4
Flask==2.1.2 
Jinja2==3.1.2 
MarkupSafe==2.1.1 
Werkzeug==2.1.2 
click==8.1.3 
waitress==2.1.2